Fix pipeline v3 xd
Should I drop Wconversions?

+++ Leetcode/1792_bis_bis.cpp
@@ -36,7 +36,7 @@ class Solution {
                              return acc + static_cast<double>(single_class[0]) /
                                               single_class[1];
                            }) /
-           classes.size();
+           static_cast<double>(classes.size());
   }
 };
